+++ pytest_dash/errors.py
@@ -1,0 +1,2 @@
+class NoAppFoundError(Exception):
+    """No `app` was found in the file."""
